+++ tests/system/app/tDataVisualization.m
@@ -1,22 +1,9 @@
-classdef tDataVisualization < matlab.uitest.TestCase
+classdef tDataVisualization < AppTestCase
 % TDATAVISUALIZATION System tests for "DataVisualization" app.
 
     properties (TestParameter)
         ValidMission = {"Bartington", "HelioSwarm", "IMAP"}
-        InvalidMission = {"Solar Orbiter", "Not a Mission"}
-    end
-
-    methods (TestClassSetup)
-
-        % Skip tests on GitHub CI runner.
-        function skipOnGitHub(testCase)
-            testCase.assumeTrue(isempty(getenv("GITHUB_ACTIONS")), "Tests cannot run on GitHub CI runner.");
-        end
-
-        % Close all figures opened by test.
-        function closeTestFigures(testCase)
-            testCase.applyFixture(mag.test.fixture.CleanupFigures());
-        end
+        InvalidMission = {"SolarOrbiter", "Not a Mission"}
     end
 
     methods (Test)
@@ -24,16 +11,34 @@ classdef tDataVisualization < matlab.uitest.TestCase
         function startApp_validMission(testCase, ValidMission)
 
             % Exercise.
-            app = DataVisualization(ValidMission);
-            testCase.addTeardown(@() delete(app));
+            app = testCase.createAppWithCleanup(ValidMission);
 
             % Verify.
             testCase.verifySize(app, [1, 1], "App should have expected size.");
             testCase.verifyClass(app, "DataVisualization", "App should be of expected class.");
+
+            testCase.verifyAppUIElementStatus(app, "off");
         end
 
         function startApp_invalidMission(testCase, InvalidMission)
             testCase.verifyError(@() DataVisualization(InvalidMission), ?MException, "Error should be thrown on invalid mission.");
         end
+
+        % function selectMission(testCase, ValidMission)
+        % 
+        %     % Set up.
+        %     app = testCase.createAppWithCleanup(ValidMission);
+        % 
+        %     % Exercise.
+        %     testCase.press(app.ToolbarManager.MissionPushTool);
+        % 
+        %     testCase.assertNotEmpty(app.SelectMissionDialog, "Mission selection dialog should not be empty.");
+        % 
+        %     testCase.choose(app.SelectMissionDialog.MissionDropDown, ValidMission);
+        %     testCase.press(app.SelectMissionDialog.SelectButton);
+        % 
+        %     % Verify.
+        %     testCase.verifyEqual(app.Mission, ValidMission, "Mission should be selected.");
+        % end
     end
 end